move header component
